refactor: 역 이름 수정 메서드 삭제
