+++ src/main/java/subway/ui/StationController.java
@@ -1,14 +1,20 @@
 package subway.ui;
 
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-import subway.dto.StationRequest;
-import subway.dto.StationResponse;
-import subway.application.StationService;
-
 import java.net.URI;
 import java.sql.SQLException;
 import java.util.List;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import subway.application.StationService;
+import subway.ui.dto.StationRequest;
+import subway.ui.dto.StationResponse;
 
 @RestController
 @RequestMapping("/stations")
@@ -33,12 +39,6 @@ public class StationController {
     @GetMapping("/{id}")
     public ResponseEntity<StationResponse> showStation(@PathVariable Long id) {
         return ResponseEntity.ok().body(stationService.findStationResponseById(id));
-    }
-
-    @PutMapping("/{id}")
-    public ResponseEntity<Void> updateStation(@PathVariable Long id, @RequestBody StationRequest stationRequest) {
-        stationService.updateStation(id, stationRequest);
-        return ResponseEntity.ok().build();
     }
 
     @DeleteMapping("/{id}")
